Aborting undeploy if there are too much instances. Issue - https://wso2.org/jira/browse/BPS-363

+++ components/bpel/org.wso2.carbon.bpel/src/main/java/org/wso2/carbon/bpel/core/ode/integration/store/TenantProcessStoreImpl.java
@@ -36,6 +36,7 @@ import org.wso2.carbon.bpel.common.config.EndpointConfiguration;
 import org.wso2.carbon.bpel.core.BPELConstants;
 import org.wso2.carbon.bpel.core.internal.BPELServiceComponent;
 import org.wso2.carbon.bpel.core.ode.integration.BPELServerImpl;
+import org.wso2.carbon.bpel.core.ode.integration.config.BPELServerConfiguration;
 import org.wso2.carbon.bpel.core.ode.integration.config.bam.BAMServerProfile;
 import org.wso2.carbon.bpel.core.ode.integration.config.bam.BAMServerProfileBuilder;
 import org.wso2.carbon.bpel.core.ode.integration.store.clustering.BPELProcessStateChangedCommand;
@@ -288,7 +289,7 @@ public class TenantProcessStoreImpl implements TenantProcessStore {
      * @param bpelPackageName Name of the BPEL package which going to be undeployed
      */
     public void undeploy(String bpelPackageName)
-            throws RegistryException {
+            throws RegistryException, TooMuchInstancesException {
 
         if (log.isDebugEnabled()) {
             log.debug("Un-deploying BPEL package " + bpelPackageName + " ....");
@@ -329,6 +330,12 @@ public class TenantProcessStoreImpl implements TenantProcessStore {
             throw re;
         }
 
+        //check the instance count to be deleted
+        long instanceCount = getInstanceCountForPackage(versionsOfThePackage);
+        if (instanceCount > BPELServerImpl.getInstance().getBpelServerConfiguration().getBpelInstanceDeletionLimit()) {
+            throw new TooMuchInstancesException("Too much instances to delete");
+        }
+
         for (String nameWithVersion : versionsOfThePackage) {
             parentProcessStore.deleteDeploymentUnitDataFromDB(nameWithVersion);
             Utils.deleteInstances(getProcessesInPackage(nameWithVersion));
@@ -364,6 +371,14 @@ public class TenantProcessStoreImpl implements TenantProcessStore {
 //        parentProcessStore.sendProcessDeploymentNotificationsToCluster(
 //                new BPELPackageUndeployedCommand(versionsOfThePackage, bpelPackageName, tenantId),
 //                configurationContext);
+    }
+
+    private int getInstanceCountForPackage(List<String> versionsOfThePackage) {
+        int count = 0;
+        for (String versionName : versionsOfThePackage ) {
+            count += Utils.getInstanceCountForProcess(getProcessesInPackage(versionName));
+        }
+        return count;
     }
 
     /**
